+++ integration-tests/emulated/tests/bridges/bridge-hub-polkadot/src/tests/snowbridge.rs
@@ -13,12 +13,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 use crate::*;
-use asset_hub_polkadot_runtime::xcm_config::bridging::to_ethereum::{
-	BridgeHubEthereumBaseFee, EthereumNetwork,
+use asset_hub_polkadot_runtime::xcm_config::{
+	bridging::to_ethereum::{BridgeHubEthereumBaseFee, EthereumNetwork},
+	RelayTreasuryPalletAccount,
 };
 use bp_bridge_hub_polkadot::snowbridge::CreateAssetCall;
 use bridge_hub_polkadot_runtime::{
-	EthereumBeaconClient, EthereumGatewayAddress, EthereumInboundQueue, Runtime, RuntimeOrigin,
+	bridge_to_ethereum_config::EthereumGatewayAddress, EthereumBeaconClient, EthereumInboundQueue,
+	Runtime, RuntimeOrigin,
 };
 use codec::{Decode, Encode};
 use emulated_integration_tests_common::xcm_emulator::ConvertLocation;
@@ -50,8 +52,6 @@ use system_parachains_constants::polkadot::currency::UNITS;
 
 const INITIAL_FUND: u128 = 5_000_000_000 * POLKADOT_ED;
 const CHAIN_ID: u64 = 1;
-const TREASURY_ACCOUNT: [u8; 32] =
-	hex!("af3e7da28608e13e4399cc7d14a57bdb154dde5f3d546f5f293994ef36ef7f11");
 const WETH: [u8; 20] = hex!("87d1f7fdfEe7f651FaBc8bFCB6E086C278b77A7d");
 const ETHEREUM_DESTINATION_ADDRESS: [u8; 20] = hex!("44a57ee2f2FCcb85FDa2B0B18EBD0D8D2333700e");
 const GATEWAY_ADDRESS: [u8; 20] = hex!("EDa338E4dC46038493b885327842fD3E301CaB39");
@@ -439,7 +439,7 @@ fn send_weth_asset_from_asset_hub_to_ethereum() {
 
 	BridgeHubPolkadot::fund_accounts(vec![
 		(assethub_sovereign.clone(), INITIAL_FUND),
-		(TREASURY_ACCOUNT.into(), INITIAL_FUND),
+		(RelayTreasuryPalletAccount::get(), INITIAL_FUND),
 	]);
 	AssetHubPolkadot::fund_accounts(vec![
 		(AssetHubPolkadotReceiver::get(), INITIAL_FUND),
@@ -469,6 +469,7 @@ fn send_weth_asset_from_asset_hub_to_ethereum() {
 						local: (1 * UNITS / 100).into(), // 0.01 DOT
 						remote: meth(1),
 					},
+					multiplier: FixedU128::from_rational(1, 1),
 				}
 			)
 		);
@@ -499,6 +500,11 @@ fn send_weth_asset_from_asset_hub_to_ethereum() {
 				RuntimeEvent::XcmpQueue(cumulus_pallet_xcmp_queue::Event::XcmpMessageSent { .. }) => {},
 			]
 		);
+	});
+
+	// check treasury account balance on BH before
+	let treasury_account_before = BridgeHubPolkadot::execute_with(|| {
+		<<BridgeHubPolkadot as BridgeHubPolkadotPallet>::Balances as frame_support::traits::fungible::Inspect<_>>::balance(&RelayTreasuryPalletAccount::get())
 	});
 
 	AssetHubPolkadot::execute_with(|| {
@@ -568,22 +574,27 @@ fn send_weth_asset_from_asset_hub_to_ethereum() {
 				RuntimeEvent::EthereumOutboundQueue(snowbridge_pallet_outbound_queue::Event::MessageQueued {..}) => {},
 			]
 		);
+
+		// check treasury account balance on BH after (should receive some fees)
+		let treasury_account_after = <<BridgeHubPolkadot as BridgeHubPolkadotPallet>::Balances as frame_support::traits::fungible::Inspect<_>>::balance(&RelayTreasuryPalletAccount::get());
+		let local_fee = treasury_account_after - treasury_account_before;
+
 		let events = BridgeHubPolkadot::events();
 		// Check that the local fee was credited to the Snowbridge sovereign account
 		assert!(
 			events.iter().any(|event| matches!(
 				event,
 				RuntimeEvent::Balances(pallet_balances::Event::Minted { who, amount })
-					if *who == TREASURY_ACCOUNT.into() && *amount == 50710000
+					if *who == RelayTreasuryPalletAccount::get() && *amount == local_fee
 			)),
 			"Snowbridge sovereign takes local fee."
 		);
-		// Check that the remote fee was credited to the AssetHub sovereign account
+		// Check that the remote delivery fee was credited to the AssetHub sovereign account
 		assert!(
 			events.iter().any(|event| matches!(
 				event,
-				RuntimeEvent::Balances(pallet_balances::Event::Minted { who, amount })
-					if *who == assethub_sovereign && *amount == 5025000000,
+				RuntimeEvent::Balances(pallet_balances::Event::Minted { who, .. })
+					if *who == assethub_sovereign,
 			)),
 			"AssetHub sovereign takes remote fee."
 		);