+++ CHANGELOG.md
@@ -4,10 +4,16 @@ Changelog for the runtimes governed by the Polkadot Fellowship.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## [Unreleased]
 
 - Add `pallet-vesting` to Asset Hubs ([polkadot-fellows/runtimes#269](https://github.com/polkadot-fellows/runtimes/pull/269))
 - Fix Kusama Coretime launch issues: import leases and fix renewals for short leases ([polkadot-fellows/runtimes#276](https://github.com/polkadot-fellows/runtimes/pull/276))
+
+### Fixed
+
+- Bumped `bp-header-chain` dependency and added migration to Polkadot Bridge Hub to unstuck
+  bridge with Kusama. Details may be found in relevant
+  [polkadot-sdk](https://github.com/paritytech/polkadot-sdk/pull/4208) PR.
 
 ## [1.2.1] 09.04.2024