+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Changelog for the runtimes governed by the Polkadot Fellowship.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Fixed
+
+- Bumped `bp-header-chain` dependency and added migration to Polkadot Bridge Hub to unstuck
+  bridge with Kusama. Details may be found in relevant
+  [polkadot-sdk](https://github.com/paritytech/polkadot-sdk/pull/4208) PR.
+
 ## [1.2.1] 09.04.2024
 
 ### Changed